+++ crates/rust-client/src/lib.rs
@@ -217,6 +217,7 @@ pub mod utils {
 #[cfg(feature = "testing")]
 pub mod testing {
     pub use miden_objects::testing::*;
+    pub use miden_testing::*;
 
     pub use crate::test_utils::*;
 }