+++ crates/rust-client/src/rpc/mod.rs
@@ -124,7 +124,7 @@ pub trait NodeRpcClient: Send + Sync {
 
     /// Fetches note-related data for a list of [`NoteId`] using the `/GetNotesById` rpc endpoint.
     ///
-    /// For any `[miden_objects::note::NoteType::Private]` note, the return data is only the
+    /// For any [`miden_objects::note::NoteType::Private`] note, the return data is only the
     /// [`miden_objects::note::NoteMetadata`], whereas for [`miden_objects::note::NoteType::Public`]
     /// notes, the return data includes all details.
     async fn get_notes_by_id(&self, note_ids: &[NoteId]) -> Result<Vec<FetchedNote>, RpcError>;