+++ crates/web-client/js/index.js
@@ -392,6 +392,8 @@ export class WebClient {
       // If a prover is provided, serialize it and add it to the args.
       if (prover) {
         args.push(prover.serialize());
+      } else {
+        args.push(null);
       }
 
       // Always call the same worker method.
@@ -407,9 +409,11 @@ export class WebClient {
       if (!this.worker) {
         return await this.wasmWebClient.syncState();
       }
+
       const serializedSyncSummaryBytes = await this.callMethodWithWorker(
         MethodName.SYNC_STATE
       );
+
       return wasm.SyncSummary.deserialize(
         new Uint8Array(serializedSyncSummaryBytes)
       );
@@ -434,3 +438,104 @@ Object.getOwnPropertyNames(WasmWebClient).forEach((prop) => {
     WebClient[prop] = WasmWebClient[prop];
   }
 });
+
+export class MockWebClient extends WebClient {
+  constructor(seed) {
+    super(null, seed);
+  }
+
+  /**
+   * Factory method to create a WebClient with a mock chain for testing purposes.
+   *
+   * @param serializedMockChain - Serialized mock chain data (optional). Will use an empty chain if not provided.
+   * @param seed - The seed for the account (optional).
+   * @returns A promise that resolves to a MockWebClient.
+   */
+  static async createClient(serializedMockChain, seed) {
+    // Construct the instance (synchronously).
+    const instance = new MockWebClient(seed);
+
+    // Wait for the underlying wasmWebClient to be initialized.
+    await instance.wasmWebClient.createMockClient(seed, serializedMockChain);
+
+    // Wait for the worker to be ready
+    await instance.ready;
+
+    // Return a proxy that forwards missing properties to wasmWebClient.
+    return new Proxy(instance, {
+      get(target, prop, receiver) {
+        // If the property exists on the wrapper, return it.
+        if (prop in target) {
+          return Reflect.get(target, prop, receiver);
+        }
+        // Otherwise, if the wasmWebClient has it, return that.
+        if (target.wasmWebClient && prop in target.wasmWebClient) {
+          const value = target.wasmWebClient[prop];
+          if (typeof value === "function") {
+            return value.bind(target.wasmWebClient);
+          }
+          return value;
+        }
+        return undefined;
+      },
+    });
+  }
+
+  async syncState() {
+    try {
+      if (!this.worker) {
+        return await this.wasmWebClient.syncState();
+      }
+
+      let serializedMockChain = this.wasmWebClient.serializeMockChain().buffer;
+
+      const serializedSyncSummaryBytes = await this.callMethodWithWorker(
+        MethodName.SYNC_STATE_MOCK,
+        serializedMockChain
+      );
+
+      return wasm.SyncSummary.deserialize(
+        new Uint8Array(serializedSyncSummaryBytes)
+      );
+    } catch (error) {
+      console.error("INDEX.JS: Error in syncState:", error.toString());
+      throw error;
+    }
+  }
+
+  async submitTransaction(transactionResult, prover = undefined) {
+    try {
+      if (!this.worker) {
+        return await this.wasmWebClient.submitTransaction(
+          transactionResult,
+          prover
+        );
+      }
+      const serializedTransactionResult = transactionResult.serialize();
+      const args = [serializedTransactionResult];
+
+      // If a prover is provided, serialize it and add it to the args.
+      if (prover) {
+        args.push(prover.serialize());
+      } else {
+        args.push(null);
+      }
+
+      args.push(this.wasmWebClient.serializeMockChain().buffer);
+
+      // Always call the same worker method.
+      let serializedMockChain = await this.callMethodWithWorker(
+        MethodName.SUBMIT_TRANSACTION_MOCK,
+        ...args
+      );
+
+      serializedMockChain = new Uint8Array(serializedMockChain);
+
+      this.wasmWebClient = new WasmWebClient();
+      await this.wasmWebClient.createMockClient(this.seed, serializedMockChain);
+    } catch (error) {
+      console.error("INDEX.JS: Error in submitTransaction:", error.toString());
+      throw error;
+    }
+  }
+}