+++ crates/web-client/test/mocha.global.setup.mjs
@@ -124,6 +124,7 @@ before(async () => {
         TransactionScriptInputPairArray,
         Word,
         WebClient,
+        MockWebClient,
       } = await import("./index.js");
       let rpcUrl = `http://localhost:${rpcPort}`;
       let proverUrl = null;
@@ -196,6 +197,7 @@ before(async () => {
       window.TransactionScriptInputPair = TransactionScriptInputPair;
       window.TransactionScriptInputPairArray = TransactionScriptInputPairArray;
       window.WebClient = WebClient;
+      window.MockWebClient = MockWebClient;
       window.Word = Word;
 
       // Create a namespace for helper functions